Dedup index download infrastructure
